skills/version-release/tests/test_new_functions.py
"""
測試版本發布工具的新函式 - Phase 3b 實作驗證
"""

import pytest
from pathlib import Path
from unittest.mock import Mock, patch, MagicMock
import tempfile
import yaml
import sys

# 為了測試，我們需要能夠 import version_release
# 將父目錄加入路徑
sys.path.insert(0, str(Path(__file__).parent.parent / "scripts"))

# 注意：這是測試框架設置，實際實作完成後再執行
# 此處提供測試結構


class TestLoadVersionReleaseConfig:
    """測試 load_version_release_config 函式"""

    def test_config_file_not_exists(self, tmp_path):
        """場景 1.2：配置檔不存在"""
        # Given: .version-release.yaml 不存在於專案根目錄
        # When: 執行 load_version_release_config(root)
        # Then: 回傳 DEFAULT_VERSION_RELEASE_CONFIG
        pass

    def test_config_file_valid(self, tmp_path):
        """場景 1.1：配置檔存在且格式正確"""
        pass

    def test_config_file_yaml_error(self, tmp_path):
        """場景 1.3：配置檔 YAML 格式錯誤"""
        pass

    def test_config_file_partial_missing_fields(self, tmp_path):
        """場景 1.4：配置檔部分欄位缺漏"""
        pass

    def test_config_file_extra_fields(self, tmp_path):
        """場景 1.5：配置檔包含未知欄位"""
        pass


class TestGetMonorepoVersion:
    """測試 get_monorepo_version 函式"""

    def test_todolist_exists_with_version(self, tmp_path):
        """場景 2.1：todolist.yaml 存在，current_version 欄位有效"""
        pass

    def test_todolist_not_exists(self, tmp_path):
        """場景 2.2：todolist.yaml 不存在"""
        pass

    def test_todolist_missing_current_version(self, tmp_path):
        """場景 2.3：todolist.yaml 存在，但 current_version 欄位不存在"""
        pass

    def test_current_version_non_string(self, tmp_path):
        """場景 2.4：current_version 為非字串型別"""
        pass

    def test_current_version_variant_format(self, tmp_path):
        """場景 2.5：current_version 為版本風格變體"""
        pass


class TestCheckMonorepoVersionSync:
    """測試 check_monorepo_version_sync 函式"""

    def test_l1_l2_mismatch_expected(self, tmp_path):
        """場景 3.1：L1=0.1.1，L2=1.0.0+1（預期不匹配）"""
        pass

    def test_l2_greater_than_l1(self, tmp_path):
        """場景 3.2：L2 版本大於 L1（警告）"""
        pass

    def test_l2_less_than_l1(self, tmp_path):
        """場景 3.3：L2 版本小於 L1（資訊）"""
        pass

    def test_l2_equal_to_l1(self, tmp_path):
        """場景 3.4：L2 版本等於 L1（一致）"""
        pass

    def test_l2_not_exists(self, tmp_path):
        """場景 3.5：ui/pubspec.yaml 不存在"""
        pass

    def test_version_with_build_number(self, tmp_path):
        """場景 3.6：版本字串含 build number 的比對"""
        pass


class TestPrintVersionSyncReport:
    """測試 print_version_sync_report 函式"""

    def test_normal_output_three_layers(self, capsys):
        """場景 4.1：正常輸出三層版本對比"""
        pass

    def test_output_without_l2_version(self, capsys):
        """場景 4.2：無 L2 版本的輸出"""
        pass

    def test_output_with_warning(self, capsys):
        """場景 4.3：包含警告訊息的輸出"""
        pass

    def test_output_empty_messages(self, capsys):
        """場景 4.4：空的訊息清單"""
        pass


class TestIntegration:
    """整合測試"""

    def test_standard_check_flow(self, tmp_path):
        """場景 5.1：標準 check 流程 - 正常三層版本對比"""
        pass

    def test_check_without_config_file(self, tmp_path):
        """場景 5.2：.version-release.yaml 不存在時的 check"""
        pass

    def test_check_missing_todolist(self, tmp_path):
        """場景 5.3：check 時 L1 todolist.yaml 不存在"""
        pass

    def test_check_l2_greater_than_l1(self, tmp_path):
        """場景 5.4：L2 版本大於 L1 的 check"""
        pass


class TestEdgeCases:
    """邊界條件測試"""

    def test_empty_version_string(self):
        """邊界 1：version 為空字串"""
        pass

    def test_permission_denied(self, tmp_path):
        """邊界 2：檔案權限不足"""
        pass

    def test_path_is_directory(self, tmp_path):
        """邊界 3：路徑為目錄而非檔案"""
        pass

    def test_empty_yaml_file(self, tmp_path):
        """邊界 4：空的 YAML 檔案"""
        pass


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
