skills/version-release/pyproject.toml
[project]
name = "version-release"
version = "1.0.0"
description = "Version Release Tool for Monorepo"
requires-python = ">=3.10"
dependencies = ["pyyaml>=5.0"]
